scene: more rstd

+++ src/Scene/SceneWallpaper.cpp
@@ -1364,9 +1364,9 @@ void SceneRuntimeController::loadScene() {
                     SceneParseCapabilities {
                         .directional_shadow = m_render_capabilities->directional_shadow(),
                         .max_geometry_output_vertices =
-                            m_render_capabilities->max_geometry_output_vertices,
+                            u32(m_render_capabilities->max_geometry_output_vertices),
                         .max_geometry_total_output_components =
-                            m_render_capabilities->max_geometry_total_output_components,
+                            u32(m_render_capabilities->max_geometry_total_output_components),
                     },
             });
         if (parsed.is_err()) {
